+++ haraka-aesni/Makefile
@@ -3,8 +3,8 @@ CFLAGS = -Wall -Wextra -Wpedantic -O3 -std=c99 -march=native -fomit-frame-pointe
 
 THASH = robust
 
-SOURCES =          hash_haraka.c hash_harakax4.c thash_haraka_$(THASH).c thash_haraka_$(THASH)x4.c address.c randombytes.c wots.c utils.c utilsx4.c fors.c sign.c haraka.c
-HEADERS = params.h hash.h        hashx4.h        thash.h               thashx4.h               address.h randombytes.h wots.h utils.h utilsx4.h fors.h api.h haraka.h harakax4.h
+SOURCES =          hash_haraka.c hash_harakax4.c thash_haraka_$(THASH).c thash_haraka_$(THASH)x4.c address.c randombytes.c merkle.c wots.c utils.c utilsx4.c fors.c sign.c haraka.c
+HEADERS = params.h hash.h        hashx4.h        thash.h               thashx4.h               address.h randombytes.h merkle.c wots.h utils.h utilsx4.h fors.h api.h haraka.h harakax4.h
 
 DET_SOURCES = $(SOURCES:randombytes.%=rng.%)
 DET_HEADERS = $(HEADERS:randombytes.%=rng.%)
